syscalls/inotify07: test inotify watch after drop caches
Add watch on a directory and drop caches dentry and inode cache.
inotify pins the directory inode in cache, but not the dentry.
Execute operations on directory and child and expect events to be
reported on directory watch. This will fail if file system does
not obtain the pinned inode to the new allocated dentry after drop
caches.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>