syscalls/fanotify20: new test for FAN_REPORT_PIDFD
This test ensures that the interface returns the expected error status
code -EINVAL when an invalid flag is supplied alongside the new
FAN_REPORT_PIDFD initialization flag. Currently, FAN_REPORT_TID is the
only initialization flag that is not permitted in conjunction with
FAN_REPORT_PIDFD, so we explicitly cover this case.

We also add a number of other simple test cases to ensure that
initialization behavior is working as expected when other supplemental
information object related flags are provided alongside
FAN_REPORT_PIDFD.

Signed-off-by: Matthew Bobrowski <repnop@google.com>